Move SCM to managers.lifecycle package
