--- baseAndroid/src/main/java/com/smartdevicelink/proxy/SystemCapabilityManager.java
@@ -1,1 +1,0 @@
-../../../../../../../base/src/main/java/com/smartdevicelink/proxy/SystemCapabilityManager.java
